// csm.web.app/src/app/project/project-common/enums/relation-type.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2021
 *
 * *************************************************************************
 */

export enum RelationTypeEnum {
    FinishToStart = 'FINISH_TO_START',
    PartOf = 'PART_OF',
}